requirements.txt
psutil==5.9.8
pandas==2.2.2
matplotlib==3.9.0
PyYAML
